updated the maintemplate and the contact2/2
added the margin to the form main template have to do the alignment with the stepper. In contact  2/2  added the aligments and also added the icons and in the image folder the images starting form icon is related to the contact page. still there are some changes to do in it and will do