Did some edits
Made the cards scroll bar stop escaping the card. Edited the button in the textfield button display to not touch the border.Edited the text field in work experience 2 to avoid the content typed in it to overlap with the word count. Have to edit it to have the same border color as material ui textfield component when clicked on it